20231114 17:26 사용자 임의 중단 시 백그라운드 프로세스 즉시 Kill 구현 완료

+++ WinFormsAppTest/MainForm.cs
@@ -343,8 +343,15 @@ namespace WinFormsAppTest
         }
 
         //공장초기화 버튼
-        private void btn_factory_reset_Click(object sender, EventArgs e)
+        private void BtnFactoryReset_Click(object sender, EventArgs e)
         {
+            DialogResult = MessageBox.Show("설정값 관련 파일을 초기화 합니다.", "DB 초기화", MessageBoxButtons.YesNo);
+
+            if (DialogResult == DialogResult.No) 
+            {
+                return;
+            }
+
             //FactoryReset(csv_path);
             DeleteAllTables(databaseFileName, tablename);
             CreateTable(databaseFileName, tablename);
